requirements.txt
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.8.4
streamlit==1.42.0
seaborn==0.13.2
plotly==6.0.0